+++ www/api_proxy.php
@@ -143,6 +143,7 @@ $ALLOWED_PROXY_PREFIXES = [
     '/tasks/',
     '/groups',
     '/maintenance/',
+    '/approvals',
 ];
 $pathAllowed = false;
 foreach ($ALLOWED_PROXY_PREFIXES as $prefix) {
@@ -170,7 +171,7 @@ $ADMIN_ONLY_PREFIXES = [
     '/scan_server_users', '/sshd_allow_user', '/remove_user_keys', '/delete_remote_user',
     '/server_user_remove_key', '/admin/', '/policy/', '/exclude_user',
     '/server_lifecycle', '/update_security_exec', '/drift/', '/tasks/',
-    '/groups', '/maintenance/windows',
+    '/groups', '/maintenance/windows', '/approvals',
 ];
 if ($roleId < ROLE_ADMIN) {
     foreach ($ADMIN_ONLY_PREFIXES as $prefix) {